requirements.txt
vkbottle==4.6.2
python-dotenv==1.1.1
fastapi==0.118.3